move all the ff stuff to an init to share across all packages
