+++ local/unleash-compose.yaml
@@ -47,7 +47,3 @@ services:
       timeout: 1m
       retries: 5
       start_period: 10s
-
-volumes:
-  db:
-    driver: local
